[TEMPURA-2060] Add Painter image to composable Badge
