+++ app/src/main/java/net/skyscanner/backpack/demo/compose/BadgeStory.kt
@@ -29,12 +29,14 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
 import net.skyscanner.backpack.compose.badge.BpkBadge
 import net.skyscanner.backpack.compose.badge.BpkBadgeType
 import net.skyscanner.backpack.compose.icon.BpkIcon
 import net.skyscanner.backpack.compose.theme.BpkTheme
 import net.skyscanner.backpack.compose.tokens.BpkSpacing
 import net.skyscanner.backpack.compose.tokens.TickCircle
+import net.skyscanner.backpack.demo.R
 import net.skyscanner.backpack.demo.components.BadgeComponent
 import net.skyscanner.backpack.demo.meta.ComposeStory
 
@@ -95,6 +97,14 @@ private fun BadgeRow(
                 else -> BpkIcon.TickCircle
             },
         )
+
+        BpkBadge(
+            text = type.toString(),
+            icon = painterResource(id = R.drawable.sample_icon),
+            modifier = Modifier
+                .weight(1f)
+                .wrapContentWidth(align = Alignment.CenterHorizontally),
+            type = type,
+        )
     }
 }
-// }
