+++ backpack-compose/src/main/kotlin/net/skyscanner/backpack/compose/button/internal/BpkButtonImpl.kt
@@ -176,7 +176,7 @@ internal fun ButtonDrawable(
 
 private val ButtonShape = RoundedCornerShape(BpkBorderRadius.Sm)
 
-private fun Modifier.defaultIconSize(size: BpkIconSize): Modifier =
+internal fun Modifier.defaultIconSize(size: BpkIconSize): Modifier =
     when (size) {
         BpkIconSize.Small -> this.requiredSize(BpkSpacing.Base, BpkSpacing.Base)
         BpkIconSize.Large -> this.requiredSize(BpkSpacing.Lg, BpkSpacing.Lg)
